delete job reducer and integrate it into rootReducer since its only one
