+++ src/reducers/rootReducer.js
@@ -1,6 +1,40 @@
-import { combineReducers } from 'redux'
-import jobsReducer from './jobsReducer'
+const initState = {
+  jobs: [
+    {
+      "id": "d98ef1c8-0559-4784-a18c-12f0b21d0315",
+      "type": "Full Time",
+      "url": "https://jobs.github.com/positions/d98ef1c8-0559-4784-a18c-12f0b21d0315",
+      "created_at": "Mon Oct 05 16:37:44 UTC 2020",
+      "company": "Agrian Inc.",
+      "company_url": "http://www.agrian.com",
+      "location": "Fresno, CA",
+      "title": "Software Engineer - Web UI"
+    },
+    {
+      "id": "a3b5a3b3-9379-4274-a55d-ef9c1ad2578c",
+      "type": "Full Time",
+      "url": "https://jobs.github.com/positions/a3b5a3b3-9379-4274-a55d-ef9c1ad2578c",
+      "created_at": "Mon Oct 05 16:31:56 UTC 2020",
+      "company": "Agrian.com",
+      "company_url": "http://Agrian.com",
+      "location": "Fresno, CA",
+      "title": "Software Engineer - React Native"
+    },
+    {
+      "id": "421871bb-05d3-4fbe-b00c-3f372fa35584",
+      "type": "Full Time",
+      "url": "https://jobs.github.com/positions/421871bb-05d3-4fbe-b00c-3f372fa35584",
+      "created_at": "Tue Aug 04 17:23:12 UTC 2020",
+      "company": "Sticker Mule",
+      "company_url": "https://www.stickermule.com/careers",
+      "location": "Anywhere",
+      "title": "Javascript Engineer"
+    }
+  ]
+}
 
-const rootReducer= combineReducers({jobsReducer})
+const rootReducer = (state = initState, action) => {
+  return state;
+}
 
-export default rootReducer;
+export default rootReducer
